Split out generic shield alert code into a base class
