+++ Source/Common/FrontierDevelopments-Shields.csproj
@@ -66,6 +66,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Shields\Alerts\Alert_ShieldBase.cs" />
     <Compile Include="Shields\Alerts\Alert_ShieldLowPower.cs" />
     <Compile Include="Shields\Buildings\Building_ElectricShield.cs" />
     <Compile Include="Shields\Comps\Comp_ShieldBase.cs" />
