+++ Source/Common/Shields/Alerts/Alert_ShieldLowPower.cs
@@ -1,33 +1,13 @@
-using System.Collections.Generic;
 using System.Linq;
 using FrontierDevelopments.Shields.Buildings;
-using RimWorld;
 using UnityEngine;
 using Verse;
 
 namespace FrontierDevelopments.Shields.Alerts
 {
-    public class Alert_ShieldLowPower : Alert
+    public class Alert_ShieldLowPower : Alert_ShieldBase
     {
-        private List<Thing> _offenders;
-        
-        public override AlertReport GetReport()
-        {
-            _offenders = GetOffenders().ToList();
-            if (!_offenders.Any()) return false;
-            return AlertReport.CulpritsAre(_offenders);
-        }
-
-        private static IEnumerable<Thing> GetOffenders()
-        {
-            return Find.Maps
-                .SelectMany(map => map.GetComponent<ShieldManager>().Fields)
-                .SelectMany(field => field.Emitters)
-                .Where(IsOffender)
-                .Select(shield => shield.Thing);
-        }
-
-        private static bool IsOffender(IShield shield)
+        protected override bool IsOffender(IShield shield)
         {
             return shield.Status.OfType<Building_ElectricShield.ShieldStatusBatteryLow>().Any();
         }
@@ -42,7 +22,7 @@ namespace FrontierDevelopments.Shields.Alerts
         public override TaggedString GetExplanation()
         {
             return "fd.shields.alert.lowpower.description".Translate()
-                .Replace("{0}", "" + _offenders.Count)
+                .Replace("{0}", "" + Offenders.Count)
                 .Replace("{1}", "50");
         }
     }
